delete comment working !

+++ backend/controllers/posts.js
@@ -2,19 +2,25 @@ const db = require("../models");
 const Post = db.posts;
 const Comment = db.comments;
 const User = db.user;
+const jwt = require('jsonwebtoken');
 
 //CRUD
 
 //Création d'un article
-exports.createPost = (req, res, next) => {
+exports.createPost = async (req, res, next) => {
   const post = { title: req.body.title, content: req.body.content, userId: req.body.userId }
+ 
+  if (req.body.title === "" || req.body.content === "") {
+    return res.status(400).json({ error: "Merci de remplir tous les champs." });
+  }
+
   Post.create(post)
     .then(() => res.status(201).json({ message: 'Post enregistré !' }))
     .catch(error => res.status(400).json({ error }));
 };
 
 //Récupération de tous les articles
-exports.getAllPosts = (req, res, next) => {
+exports.getAllPosts = async (req, res, next) => {
   Post.findAll({ include: User.name })
     .then((posts) => {
       res.status(200).json(posts);
@@ -25,7 +31,7 @@ exports.getAllPosts = (req, res, next) => {
 };
 
 //Récupération d'un article en cherchant son id (where)
-exports.getOnePost = (req, res, next) => {
+exports.getOnePost = async (req, res, next) => {
   Post.findOne({
     where: { id: req.params.id },
     include: [
@@ -48,12 +54,19 @@ exports.getOnePost = (req, res, next) => {
 
 
 //Suppression d'un article en cherchant son id (where)
-exports.deletePost = (req, res, next) => {
-  Post.destroy({ where: { id: req.params.id } })
+
+exports.deletePost = async (req, res, next) => {
+  const userId = req.params.userId
+  if(Post.userId === userId || User.isAdmin === true) {
+    Post.destroy({ where: { id: req.params.id } })
     .then((post) => {
       res.status(200).json(post);
     })
     .catch((error) => {
       res.status(404).json({ error: error });
     });
+  }
+  else {
+    console.log(error)
+  }
 };
